Switched to Uint8Array so works with real MIDI data

+++ test/resources/test-utils.js
@@ -1,11 +1,11 @@
 function testUtils() {
 	function ab2str(buf) {
-		return String.fromCharCode.apply(null, new Uint16Array(buf));
+		return String.fromCharCode.apply(null, new Uint8Array(buf));
 	}
 
 	function str2ab(str) {
 		var buf = new ArrayBuffer(str.length * 2),
-			bufView = new Uint16Array(buf);
+			bufView = new Uint8Array(buf);
 
 		for (var i = 0, strLen = str.length; i < strLen; i++) {
 			bufView[i] = str.charCodeAt(i);
